app/requirements.txt
pydantic
psycopg2-binary
uvicorn
pycryptodome
fastapi
APscheduler
requests